requirements.txt
django==3.1.7
djangorestframework
mysqlclient